seperate ranking logic from API endpoint into a service.
