+++ web/app/api/rankings.py
@@ -1,193 +1,17 @@
 import docker
-import json
 import time
-from datetime import datetime
-import requests
 from flask import jsonify, request, current_app
 from app import api
 from app.models import db, Session, System, Result, Feedback
-from app.services.interleave_service import tdi
+from app.services.interleave_service import interleave_rankings
 from app.utils import create_dict_response
 from pytz import timezone
 from . import api
 from app.services.session_service import create_new_session
+from app.services.ranking_service import query_system
 
 client = docker.DockerClient(base_url="unix://var/run/docker.sock")
 tz = timezone("Europe/Berlin")
-
-
-def single_ranking(ranking):
-    """
-    Add ranking to database
-
-    @param ranking: a ranking of type Result
-
-    @return
-        ranked items (dict)
-
-    """
-    db.session.add(ranking)
-    db.session.commit()
-
-    return ranking.items
-
-
-def interleave(ranking_exp, ranking_base):
-    """
-    Create interleaved ranking from experimental and baseline system
-    Used method: Team-Draft-Interleaving (TDI) [1]
-
-    [1] "How Does Clickthrough Data Reflect Retrieval Quality?"
-        Radlinski, Kurup, Joachims
-        Published in CIKM '15 2015
-
-    @param ranking_exp:     experimental ranking (Result)
-    @param ranking_base:    baseline ranking (Result)
-    @return:                interleaved ranking (dict)
-    """
-    base = {k: v.get("docid") for k, v in ranking_base.items.items()}
-    exp = {k: v.get("docid") for k, v in ranking_exp.items.items()}
-
-    item_dict = tdi(base, exp)
-    ranking = Result(
-        session_id=ranking_exp.session_id,
-        system_id=ranking_exp.system_id,
-        type="RANK",
-        q=ranking_exp.q,
-        q_date=ranking_exp.q_date,
-        q_time=ranking_exp.q_time,
-        num_found=ranking_exp.num_found,
-        hits=ranking_base.num_found,
-        page=ranking_exp.page,
-        rpp=ranking_exp.rpp,
-        items=item_dict,
-    )
-
-    db.session.add(ranking)
-    db.session.commit()
-
-    ranking_id = ranking.id
-    ranking.tdi = ranking_id
-
-    ranking_exp.tdi = ranking_id
-    db.session.add(ranking_exp)
-    db.session.commit()
-
-    ranking_base.tdi = ranking_id
-    db.session.add(ranking_base)
-    db.session.commit()
-
-    return ranking.items
-
-
-def rest(container_name, query, rpp, page):
-    """
-    Produce container-ranking via rest-call implementation
-
-    @param container_name:  container name (str)
-    @param query:           search query (str)
-    @param rpp:             results per page (int)
-    @param page:            page number (int)
-
-    @return:                container-ranking (dict)
-    """
-    if current_app.config["DEBUG"]:
-        container = client.containers.get(container_name)
-        ip_address = container.attrs["NetworkSettings"]["Networks"][
-            "stella-app_default"
-        ]["IPAddress"]
-        content = requests.get(
-            "http://" + ip_address + ":5000/ranking",
-            params={"query": query, "rpp": rpp, "page": page},
-        ).content
-        return json.loads(content)
-
-    content = requests.get(
-        f"http://{container_name}:5000/ranking",
-        params={"query": query, "rpp": rpp, "page": page},
-    ).content
-    return json.loads(content)
-
-
-def cmd(container_name, query, rpp, page):
-    """
-    Produce container-ranking via classical cmd-call (fallback solution, much slower than rest-implementation)
-
-    @param container_name:  container name (str)
-    @param query:           search query (str)
-    @param rpp:             results per page (int)
-    @param page:            page number (int)
-
-    @return:                container-ranking (dict)
-    """
-    container = client.containers.get(container_name)
-    cmd = "python3 /script/ranking {} {} {}".format(query, rpp, page)
-    exec_res = container.exec_run(cmd)
-    result = json.loads(exec_res.output.decode("utf-8"))
-    return result
-
-
-def query_system(container_name, query, rpp, page, session_id, type="EXP"):
-    """
-    Produce ranking from dockerized system
-
-    @param container_name:  container name (str)
-    @param query:           search query (str)
-    @param rpp:             results per page (int)
-    @param page:            page number (int)
-    @param session_id:      session_id (int)
-    @param type:            ranking type (str 'EXP' or 'BASE')
-
-    @return:                ranking (Result)
-    """
-    current_app.logger.debug(f'produce ranking with container: "{container_name}"...')
-
-    q_date = datetime.now(tz).replace(tzinfo=None, microsecond=0)
-
-    ts_start = time.time()
-    ts = round(ts_start * 1000)
-
-    # increase counter before actual request, in case of a failure
-    system = System.query.filter_by(name=container_name).first()
-    if query in current_app.config["HEAD_QUERIES"]:
-        system.num_requests += 1
-    else:
-        system.num_requests_no_head += 1
-    db.session.commit()
-
-    if current_app.config["REST_QUERY"]:
-        result = rest(container_name, query, rpp, page)
-    else:
-        result = cmd(container_name, query, rpp, page)
-
-    ts_end = time.time()
-    # calc query execution time in ms
-    q_time = round((ts_end - ts_start) * 1000)
-
-    item_dict = {
-        i + 1: {"docid": result["itemlist"][i], "type": type}
-        for i in range(0, len(result["itemlist"]))
-    }
-
-    ranking = Result(
-        session_id=session_id,
-        system_id=System.query.filter_by(name=container_name).first().id,
-        type="RANK",
-        q=query,
-        q_date=q_date,
-        q_time=q_time,
-        num_found=result["num_found"],
-        page=page,
-        rpp=rpp,
-        items=item_dict,
-    )
-
-    # system = System.query.filter_by(name=container_name).first()
-    # system.num_requests += 1
-    db.session.add(ranking)
-    db.session.commit()
-
-    return ranking
 
 
 @api.route("/test/<string:container_name>", methods=["GET"])
@@ -373,7 +197,7 @@ def ranking():
     )
 
     if current_app.config["INTERLEAVE"]:
-        response = interleave(ranking_exp, ranking_base)
+        response = interleave_rankings(ranking_exp, ranking_base)
         response_complete = {
             "header": {
                 "sid": ranking_exp.session_id,
@@ -390,7 +214,6 @@ def ranking():
             "body": response,
         }
     else:
-        response = single_ranking(ranking_exp)
         response_complete = {
             "header": {
                 "sid": ranking_exp.session_id,
@@ -401,7 +224,7 @@ def ranking():
                 "hits": ranking_exp.num_found,
                 "container": {"exp": container_name},
             },
-            "body": response,
+            "body": ranking_exp.items,
         }
 
     return jsonify(response_complete)
